Refactor contrast and manual tracking gui into a single module
