+++ alignment-software/alignment_software/gui/window_main.py
@@ -4,9 +4,9 @@ from tkinter.messagebox import showwarning
 from alignment_software.engine.particle_tracking import (
     ParticlePositionContainer
 )
-from .manual_track.step_manual_track import ManualTrackStep
+from .manual_track import ManualTrackStep
 from .common import AsyncHandler
-from .contrast.step_contrast import ContrastStep
+from .contrast import ContrastStep
 from .loading import LoadingStep
 from .transform import TransformStep
 from .coarse_align.step_coarse_align import CoarseAlignStep
